fix: timer bug
