+++ src/components/editor/plugins/link/insert/index.jsx
@@ -1,0 +1,12 @@
+import { createCommand } from "lexical";
+import { forwardRef } from "react";
+
+export const LINK_INSERT_COMMAND = createCommand();
+
+const LinkInsertToolbar = forwardRef(({ pos, isOpen }, ref) => {
+
+})
+
+export default function LinkInsertPlugin() {
+
+}
